requirements.txt
python-dotenv==0.12.0
typing==3.7.4.1
python-telegram-bot == 11.1.0
